refactor: simplify writer, schema evolution and generated columns
